refactored some methods to utils.py for better reusability
